requirements.txt
python-telegram-bot~=13.15
alembic==1.9.2
click~=8.1.3
psycopg2~=2.9.5
schedule~=1.1.0
SQLAlchemy~=1.4.44